Added scripts to create event list and dl3 index fits files
