+++ lstchain/data/data_selection_cuts.yml
@@ -1,0 +1,14 @@
+general:
+  gammaness: 0.8
+  fov_offset: 2.83
+
+data:
+ intensity: 200
+ r: 1
+ wl: 0.1
+ leakage2_intensity: 0.2 # pre lstchain v0.6
+ leakage_intensity_width_2: 0.2 # since lstchain v0.6
+ alpha: 8
+
+mc:
+  xi_cut: 0.3
