+++ lstchain/data/pyirf_config.yml
@@ -1,0 +1,83 @@
+general:
+ output_table_name: 'table_best_cutoff'
+ template_input_file: 'dl2_gamma_updated.h5'
+ outdir: ''
+ indir: ''
+
+analysis:
+
+ # Theta square cut optimisation (opti, fixed, r68)
+ thsq_opt:
+  type: 'r68'
+  value: 0.2  # In degree, necessary for type fixed
+
+ # Observation time to estimate best cuts corresponding to best sensitivity
+ obs_time:
+   value: 50
+   unit: 'h'
+ # Normalisation between ON and OFF regions
+ alpha: 0.2
+ # Minimimal significance
+ min_sigma: 5
+ # Minimal number of gamma-ray-like
+ min_excess: 10
+ # Minimal fraction of background events for excess comparison
+ bkg_syst: 0.05
+
+ # Reco energy binning
+ ereco_binning:  # TeV
+  emin: 0.05
+  emax: 50
+  nbin: 21
+
+ # Reco energy binning
+ etrue_binning:  # TeV
+  emin: 0.05
+  emax: 50
+  nbin: 10
+
+# Energy migration binning
+ emigra_binning:
+  nbin: 120
+
+particle_information:
+ gamma:
+  n_events_per_file: 22500000
+  n_files: 1
+  e_min: 0.05
+  e_max: 50
+  gen_radius: 1000
+  diff_cone: 0
+  gen_gamma: 2
+
+ proton:
+  n_events_per_file: 22500000
+  n_files: 1
+  e_min: 0.05
+  e_max: 500
+  gen_radius: 1000
+  diff_cone: 10
+  gen_gamma: 2
+  offset_cut: 1
+
+ electron:
+  n_events_per_file: 22500000
+  n_files: 1
+  e_min: 0.05
+  e_max: 300
+  gen_radius: 1000
+  diff_cone: 10
+  gen_gamma: 2
+  offset_cut: 1
+
+
+column_definition:
+ # Column name for true energy
+ mc_energy: 'mc_energy'
+ # Column name for reconstructed energy
+ reco_energy: 'reco_energy'
+ # Column name for classification output
+ classification_output:
+  name: 'gammaness'
+  range: [0, 1]
+ angular_distance_to_the_src: 'xi'
