+++ lstchain/hdu/__init__.py
@@ -1,0 +1,11 @@
+from .hdu_table import (
+    make_hdu,
+    create_obs_hdu_index,
+    create_event_list,
+    )
+
+__all__ = [
+    'make_hdu',
+    'create_obs_hdu_index',
+    'create_event_list'
+    ]
